PD-597: [WIP] Pull triggers into org and projcet selects, tests failing
