+++ packages/mongo-nav/src/mongo-select/ProjectSelect.tsx
@@ -1,30 +1,46 @@
 import React, { useState, useEffect } from 'react';
+
+// leafygreen-ui
+import Icon from '@leafygreen-ui/icon';
 import Button from '@leafygreen-ui/button';
-import Input from './Input';
 import { css, cx } from '@leafygreen-ui/emotion';
-import ProjectTrigger from './ProjectTrigger';
-import { useOnElementClick } from '../on-element-click-provider';
-import { onKeyDown } from './selectHelpers';
+import { createDataProp } from '@leafygreen-ui/lib';
 import {
   Menu,
   FocusableMenuItem,
   MenuItem,
   MenuSeparator,
 } from '@leafygreen-ui/menu';
+
+// mongo-nav
+import { facepaint } from '../breakpoints';
+import { InteractionRingWrapper } from '../helpers';
+import { useOnElementClick } from '../on-element-click-provider';
+import { textLoadingStyle, iconLoadingStyle } from '../styles';
+import {
+  CurrentProjectInterface,
+  MongoNavInterface,
+  NavElement,
+  ProjectInterface,
+} from '../types';
+
+// mongo-select
+import Input from './Input';
+import { onKeyDown } from './selectHelpers';
 import { BaseMongoSelectProps } from './types';
 import {
-  ProjectInterface,
-  CurrentProjectInterface,
-  NavElement,
-  MongoNavInterface,
-} from '../types';
-import {
+  activeButtonStyle,
+  baseButtonStyle,
+  caretBaseStyle,
+  iconColorStyle,
   menuContainerStyle,
   menuItemContainerStyle,
-  ulStyle,
   nameStyle,
+  selectedStyle,
+  ulStyle,
 } from './styles';
 
+// styles
 const projectButtonStyle = css`
   display: flex;
   justify-content: space-around;
@@ -32,21 +48,57 @@ const projectButtonStyle = css`
   padding-bottom: 16px;
 `;
 
+const projectTriggerStyle = css`
+  justify-content: space-around;
+  border-color: transparent;
+  border-radius: 5px;
+  padding: 2px;
+  width: 174px;
+  height: 28px;
+
+  ${facepaint({
+    width: ['196px', '106px', '106px'],
+    height: ['28px', '36px', '36px'],
+  })}
+
+  &:focus {
+    outline: none;
+  }
+`;
+
+const projectTriggerWrapperStyle = css`
+  margin-left: 16px;
+  margin-right: 2px;
+  z-index: 1;
+
+  ${facepaint({
+    marginLeft: ['16px', '0px', '16px'],
+  })}
+`;
+
+const projectTriggerRingStyle = css`
+  border-radius: 7px;
+`;
+
+// types
 interface ProjectMongoSelectProps extends BaseMongoSelectProps {
   data?: Array<ProjectInterface>;
   current?: CurrentProjectInterface;
   constructProjectURL: NonNullable<MongoNavInterface['constructProjectURL']>;
 }
 
+// data props
+const projectTriggerDataProp = createDataProp('project-trigger');
+
 function ProjectSelect({
   current,
   data = [],
-  onClick,
   onChange: onChangeProp,
   constructProjectURL,
   urls = {},
-  loading,
+  loading = false,
   className,
+  ...rest
 }: ProjectMongoSelectProps) {
   const [value, setValue] = useState('');
   const [open, setOpen] = useState(false);
@@ -57,8 +109,6 @@ function ProjectSelect({
       setValue('');
     }
   }, [open]);
-
-  let renderedData = data;
 
   const filterData = () => {
     const sanitizedValue = value.replace(/\\/g, '\\\\');
@@ -71,18 +121,12 @@ function ProjectSelect({
     return filtered;
   };
 
-  if (!onChangeProp) {
-    renderedData = filterData();
-  }
+  const renderedData = onChangeProp ? data : filterData();
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value;
     setValue(value);
-
-    if (onChangeProp) {
-      onChangeProp(value, e);
-      return;
-    }
+    onChangeProp?.(value, e);
   };
 
   const renderProjectOption = (datum: ProjectInterface) => {
@@ -95,7 +139,6 @@ function ProjectSelect({
         key={projectId}
         active={isActive}
         className={cx(menuItemContainerStyle, nameStyle)}
-        onClick={onClick}
         href={constructProjectURL(datum)}
       >
         {projectName} {isActive && '(current)'}
@@ -103,67 +146,101 @@ function ProjectSelect({
     );
   };
 
+  const onClickTrigger = onElementClick(
+    NavElement.ProjectNavProjectSelectTrigger,
+    () => setOpen(curr => !curr),
+  );
+
   return (
-    <ProjectTrigger
-      placeholder={current?.projectName ?? ''}
-      open={open}
-      loading={loading}
-      className={className}
-      onClick={onElementClick(NavElement.ProjectNavProjectSelectTrigger, () =>
-        setOpen(curr => !curr),
-      )}
+    <InteractionRingWrapper
+      selector={projectTriggerDataProp.selector}
+      className={projectTriggerWrapperStyle}
+      ringClassName={projectTriggerRingStyle}
     >
-      <Menu
-        usePortal={false}
-        className={menuContainerStyle}
-        justify="start"
-        spacing={0}
-        open={open}
-        setOpen={setOpen}
-        data-testid="project-select-project-list"
+      <button
+        {...rest}
+        {...projectTriggerDataProp.prop}
+        onClick={onClickTrigger}
+        data-testid="project-select-trigger"
+        className={cx(
+          baseButtonStyle,
+          projectTriggerStyle,
+          {
+            [activeButtonStyle]: open,
+            [textLoadingStyle]: loading,
+          },
+          className,
+        )}
+        disabled={loading}
+        aria-disabled={loading}
       >
-        <FocusableMenuItem>
-          <Input
-            data-testid="project-filter-input"
-            onChange={onChange}
-            onKeyDown={(e: React.KeyboardEvent) => onKeyDown(e, setValue)}
-            variant="project"
-            value={value}
-          />
-        </FocusableMenuItem>
-
-        <ul className={ulStyle}>
-          {renderedData?.map(datum => renderProjectOption(datum))}
-        </ul>
-
-        <MenuSeparator />
-
-        <li
-          onKeyDown={(e: React.KeyboardEvent) => onKeyDown(e, setValue)}
-          role="none"
-          className={projectButtonStyle}
+        <Icon
+          glyph="Folder"
+          className={cx(iconColorStyle, { [iconLoadingStyle]: loading })}
+        />
+        <span
+          className={cx(selectedStyle, { [textLoadingStyle]: loading })}
+          data-testid="project-select-active-project"
+        >
+          {current?.projectName ?? ''}
+        </span>
+        <Icon
+          size="small"
+          glyph={open ? 'CaretUp' : 'CaretDown'}
+          className={cx(caretBaseStyle, { [iconLoadingStyle]: loading })}
+        />
+        <Menu
+          usePortal={false}
+          className={menuContainerStyle}
+          justify="start"
+          spacing={0}
+          open={open}
+          setOpen={setOpen}
+          data-testid="project-select-project-list"
         >
           <FocusableMenuItem>
-            <Button
-              href={urls.viewAllProjects as string}
-              data-testid="project-select-view-all-projects"
-              onClick={onElementClick(NavElement.ProjectNavViewAllProjects)}
-            >
-              View All Projects
-            </Button>
+            <Input
+              data-testid="project-filter-input"
+              onChange={onChange}
+              onKeyDown={(e: React.KeyboardEvent) => onKeyDown(e, setValue)}
+              variant="project"
+              value={value}
+            />
           </FocusableMenuItem>
-          <FocusableMenuItem>
-            <Button
-              href={urls.newProject as string}
-              data-testid="project-select-add-new-project"
-              onClick={onElementClick(NavElement.ProjectNavAddProject)}
-            >
-              + New Project
-            </Button>
-          </FocusableMenuItem>
-        </li>
-      </Menu>
-    </ProjectTrigger>
+
+          <ul className={ulStyle}>
+            {renderedData?.map(datum => renderProjectOption(datum))}
+          </ul>
+
+          <MenuSeparator />
+
+          <li
+            onKeyDown={(e: React.KeyboardEvent) => onKeyDown(e, setValue)}
+            role="none"
+            className={projectButtonStyle}
+          >
+            <FocusableMenuItem>
+              <Button
+                href={urls.viewAllProjects as string}
+                data-testid="project-select-view-all-projects"
+                onClick={onElementClick(NavElement.ProjectNavViewAllProjects)}
+              >
+                View All Projects
+              </Button>
+            </FocusableMenuItem>
+            <FocusableMenuItem>
+              <Button
+                href={urls.newProject as string}
+                data-testid="project-select-add-new-project"
+                onClick={onElementClick(NavElement.ProjectNavAddProject)}
+              >
+                + New Project
+              </Button>
+            </FocusableMenuItem>
+          </li>
+        </Menu>
+      </button>
+    </InteractionRingWrapper>
   );
 }
